fix(ai): route auto-detected network devices to raw PTY exec (#2367)
AI exec paths (Catty, worker, MCP) decide whether to send commands raw
(no shell wrapping) based on meta.deviceType === "network". That field
only reflected the manual Network Device Mode toggle (host.deviceType),
so auto-detected switches (Huawei VRP, Cisco IOS, …) still went through
the POSIX shell wrapper. Their vendor CLIs reject the wrapped command,
the start marker never arrives, and the call stalls until the startup
timeout.

Report deviceType: "network" when the detected distro/vendor classifies
as a network device too, mirroring the terminal's own gating. This also
gives the AI system prompt vendor-CLI guidance before the user manually
flips the toggle. Mosh/ET sessions stay suppressed since they require a
shell-backed PTY.

Extract the pure buildAITerminalSessionInfo helper into its own module
for testability and add unit coverage.
