+++ components/terminalLayer/buildAITerminalSessionInfo.test.ts
@@ -1,0 +1,71 @@
+import test from "node:test";
+import assert from "node:assert/strict";
+
+import { buildAITerminalSessionInfo } from "./buildAITerminalSessionInfo";
+import type { Host, TerminalSession } from "../../types";
+
+const baseHost = (overrides: Partial<Host> = {}): Host =>
+  ({
+    id: "h1",
+    label: "sw1",
+    hostname: "10.0.0.1",
+    username: "admin",
+    protocol: "ssh",
+    ...overrides,
+  } as Host);
+
+const baseSession = (overrides: Partial<TerminalSession> = {}): TerminalSession =>
+  ({
+    id: "s1",
+    hostId: "h1",
+    protocol: "ssh",
+    status: "connected",
+    ...overrides,
+  } as TerminalSession);
+
+test("keeps explicit network deviceType", () => {
+  const info = buildAITerminalSessionInfo(
+    baseSession(),
+    baseHost({ deviceType: "network" }),
+    "linux",
+  );
+  assert.equal(info.deviceType, "network");
+});
+
+test("reports 'network' when the detected distro classifies as a network device (#2367)", () => {
+  // Huawei VRP is a known network-device vendor id; the user has NOT flipped
+  // Network Device Mode, so host.deviceType is unset.
+  const info = buildAITerminalSessionInfo(
+    baseSession(),
+    baseHost({ distro: "huawei" }),
+    "linux",
+  );
+  assert.equal(info.deviceType, "network");
+});
+
+test("does not force network for a normal linux distro", () => {
+  const info = buildAITerminalSessionInfo(
+    baseSession(),
+    baseHost({ distro: "ubuntu" }),
+    "linux",
+  );
+  assert.notEqual(info.deviceType, "network");
+});
+
+test("suppresses network deviceType for Mosh sessions", () => {
+  const info = buildAITerminalSessionInfo(
+    baseSession({ moshEnabled: true }),
+    baseHost({ distro: "huawei", deviceType: "network" }),
+    "linux",
+  );
+  assert.equal(info.deviceType, undefined);
+});
+
+test("suppresses network deviceType for ET sessions", () => {
+  const info = buildAITerminalSessionInfo(
+    baseSession({ etEnabled: true }),
+    baseHost({ deviceType: "network" }),
+    "linux",
+  );
+  assert.equal(info.deviceType, undefined);
+});
